sneakily integrated some of Liam's classes. short story: gameView now uses some of gameWorld classes
long story:
the DEMOOBJECT class is now basically acting as a wrapper for one of
liams GameMatter objects. The current setup is convenient in terms of
compatibility, and as the GameMatter develops further, I will be able to
remove layers of abstraction (get rid of the DEMO classes, and use the
full classes from the gameWorld class instead).